requirements.txt
flask>=2.2.5
openai>=0.27.0
python-dotenv>=1.0.0
